--- data.json
@@ -1,44 +1,0 @@
-{
-  "userList" : [
-    { 
-    "userId" : 1,
-    "userEmail" : "hobit22@naver.com",
-    "phoneNumber" : "01048067403",
-    "userName" : "김호빈",
-    "createdAt" : "2022-04-22",
-    "userLevel" : 20
-    },
-    {
-    "userId" : 2,
-    "userEmail" : "kt03170@gmail.com",
-    "phoneNumber" : "01025507827",
-    "userName" : "이경태",
-    "createdAt" : "2022-04-28",
-    "userLevel" : 20
-    },
-    { 
-    "userId" : 3,
-    "userEmail" : "adsxxx@naver.com",
-    "phoneNumber" : "01048067403",
-    "userName" : "짱구",
-    "createdAt" : "2022-06-22",
-    "userLevel" : 10
-    },
-    { 
-    "userId" : 4,
-    "userEmail" : "bsgagasd@naver.com",
-    "phoneNumber" : "01048067403",
-    "userName" : "공룡",
-    "createdAt" : "1022-04-22",
-    "userLevel" : 20
-    },
-    { 
-    "userId" : 5,
-    "userEmail" : "hhttjjj@naver.com",
-    "phoneNumber" : "01048067403",
-    "userName" : "거북이",
-    "createdAt" : "2022-04-23",
-    "userLevel" : 10
-    }
-    ]
-  }
