+++ src/redux/modules/teamBoard.js
@@ -248,6 +248,8 @@ export default handleActions(
     }),
     [CHECK_TODO_LIST]: (state, action) =>
       produce(state, (draft) => {
+        console.log(state)
+        console.log(action.payload)
       draft.teamBoard.toDoList = state.teamBoard.toDoList.map((e) => {
         if (e.todoId === action.payload.todoId.todoId){
           return {...e, todoCheck : action.payload.todoId.todoCheck}
